Move shared storage mapping path attributes to aws-parallelcluster-config
This attribute is used in fs_update and fetch_config recipe.
