+++ cookbooks/aws-parallelcluster-config/attributes/default.rb
@@ -5,3 +5,6 @@ default['cluster']['exported_raid_shared_dir'] = node['cluster']['raid_shared_di
 default['cluster']['exported_intel_dir'] = '/opt/intel'
 
 default['cluster']['scheduler_slots'] = 'vcpus'
+
+default['cluster']['shared_storages_mapping_path'] = "/etc/parallelcluster/shared_storages_data.yaml"
+default['cluster']['previous_shared_storages_mapping_path'] = "/etc/parallelcluster/previous_shared_storages_data.yaml"
