+++ cookbooks/aws-parallelcluster-config/libraries/helpers.rb
@@ -28,3 +28,15 @@ def is_service_installed?(service, platform_families = node['platform_family'])
     false
   end
 end
+
+# load shared storages data into node object
+def load_shared_storages_mapping
+  ruby_block "load shared storages mapping during cluster update" do
+    block do
+      require 'yaml'
+      # regenerate the shared storages mapping file after update
+      node.default['cluster']['shared_storages_mapping'] = YAML.safe_load(File.read(node['cluster']['previous_shared_storages_mapping_path']))
+      node.default['cluster']['update_shared_storages_mapping'] = YAML.safe_load(File.read(node['cluster']['shared_storages_mapping_path']))
+    end
+  end
+end
