+++ libraries/helpers.rb
@@ -350,18 +350,6 @@ def is_slurm_database_updated?
   config["Scheduling"]["SlurmSettings"]["Database"] != previous_config["Scheduling"]["SlurmSettings"]["Database"]
 end
 
-# load shared storages data into node object
-def load_shared_storages_mapping
-  ruby_block "load shared storages mapping during cluster update" do
-    block do
-      require 'yaml'
-      # regenerate the shared storages mapping file after update
-      node.default['cluster']['shared_storages_mapping'] = YAML.safe_load(File.read(node['cluster']['previous_shared_storages_mapping_path']))
-      node.default['cluster']['update_shared_storages_mapping'] = YAML.safe_load(File.read(node['cluster']['shared_storages_mapping_path']))
-    end
-  end
-end
-
 # Parse an ARN.
 # ARN format: arn:PARTITION:SERVICE:REGION:ACCOUNT_ID:RESOURCE.
 # ARN examples:
